+++ ethereum-types/src/lib.rs
@@ -8,15 +8,12 @@
 
 #![cfg_attr(not(feature = "std"), no_std)]
 
-#[macro_use]
-extern crate alloc;
-
 mod hash;
 mod uint;
 
 pub use ethbloom::{Bloom, BloomRef, Input as BloomInput};
 pub use hash::{BigEndianHash, H128, H160, H256, H264, H32, H512, H520, H64};
-pub use uint::{FromDecStrErr, U128, U256, U512, U64};
+pub use uint::{FromDecStrErr, FromStrRadixErr, FromStrRadixErrKind, U128, U256, U512, U64};
 
 pub type Address = H160;
 pub type Secret = H256;